pyproject.toml
[project]
name = "b24-transfer-lead"
version = "0.1.0"
description = "Bitrix24 Lead Transfer Application"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.12",
    "fast-bitrix24>=1.8.11",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.1",
]
